Added HashedString32 and associated tests. Accidentally included a few minor touchups from the spinlock troubleshooting as well.
Compiled using Mezz_Release on i7 1.6~2.0ghz with 16gb RAM on Ubuntu 13.10x64:

Benchmarking creation of 1000000 Strings and Hashed Strings.
To defeat optimization the String in the array is: Padding To Make Strings bigger1
To defeat optimization the Hashed String in the array is: <HashedString32 Version="1" Hash="-2147478734" String="Padding To Make Strings bigger557401" />
String sort took 1329875 Total microseconds. With 800197 in data creation and 529678 in the actual sorting.
Hash sort took 1365634 Total microseconds. With 852080 in data creation and 513554 in the actual sorting.

Randomizing sort results from previous benchamark and using for set insertion benchmark.
To defeat overoptimizing the String in the set is: Padding To Make Strings bigger1
To defeat overoptimizing the Hashes in the set is: <HashedString32 Version="1" Hash="-2147478734" String="Padding To Make Strings bigger557401" />
String sort by set insertion took (in microseconds): 1276153
Hashed sort by set insertion took (in microseconds): 674034