Made a number of changes to the Graphics::Image class.
Created a blank constructor, renamed the buffer loading methods to
"InitializeImage", to not confuse what they do with the methods that
perform actual I/O.  Renamed the former "explicit" I/O methods to just
be load/save with an underscore to note that is not the recommended use
in most cases.  Ensured there were loading and saving methods that could
be used without the presence of the Mezzanine Resource system.