Created a couple of classes for generating unique IDs.  Also re-arranged the main Mezznine header.
Created an Interval class to store a range of numbers.  Created a
UIDGenerator class that uses the Interval class to store a range of
available IDs and updates internally as appropriate to ensure it issues
unique IDs, always assigning the lowest available value.  The Mezzanine
header had a mess of "#ifndef SWIG" include portions.  Moved the headers
that SWIG doesn't like to it's own section and other headers separate,
both in alphabetical order.