Obliterated the CameraManager and moved Cameras to the SceneManager.
Also as a part of this task some of the code for the CameraController
and RayQueryTool have been cleaned up / documented.  In both cases the
camera/world can be set after creation but with an expectation that it
will be set before use (or else seg faults).  Managers that create world
proxies were given an additional method to retrieve the n-th proxy of a
certain type from it's manager, to ease situations like retrieving the
one camera that was set in a scene.  Some SoundScapeManager methods were
renamed to match counterparts in other managers.