Code styling updates and a number of touchups to the Entresol.
Added a Deinitialize method to the Entresol with some logging.  Fixed a
bug where deinit logging wasn't being committed to the log.  Added
convenience types to the Entresol for containers and iterators that
didn't have them yet.  Renamed the EngineInit method and removed the
option to immediately start the main loop.