Removed an error enum and added methods needed for sockets on Windows.
The ProcessLimit error doesn't appear to exist on Ubuntu, and it's a
very specific error that is quite unlikely to be encountered during use
of Sockets.  So much so that it's been commented out and may get removed
entirely.  Winsock in particular requires startup and cleanup methods be
called in order to use Winsock methods.  So added static methods that
wrap these.  On non-Winsock platforms it just returns success.