Code styling updates and a major bugfix in the audio subsystem.
Discovered and resolved an issue where two Audio::Sound instances using
the same audio stream would fail to playback, or the second Audio::Sound
instance playback would fail.  This was due to poor stream and buffer
management and has been corrected.  Uses of resource streams in decoders
have largely been re-written to use the standard stream API instead of
the Mezzanine API.  This is to facilitate the eventual plan to have
everything in the Mezzanine core that isn't the resource system be able
to accept standard streams for modular use outside of the engine.