Implemented more of the SubMesh wrapper.  Also created a MeshInfo class for getting at information about Meshes.
We have used an internal set of tools/code to get at mesh information in
the past and they do still exist, but will be phased out since this API
performs the same function and is more accessible.  The MeshInfo class
is minimalist and could use a little bit of expanding in the future.