Changed how Header Fields are stored in HTTPMessage.
Before we lower-cased the Header names as per the RFC on HTTP.  However
some web software (*cough* PHP *cough*) doesn't play nice with this.  So
now we store in a case sensitive way, but compare for unique-ness in a
case in-sensitive way.  Incidentally this also fixes the UnitTests
related to HTTP Requests.