Refactored the Timer class and made some documentation fixes.
The timer class is now just one class that is more lean than it used to
be.  Critical functionality is all still there though.  Updated various
parts of the engine that were using timers based on the API changes.
Along the way I noticed a bunch of documentation for private members of
classes in the Input namespace was missing, so those got documented.