Fixed an issue where CMake would fail for the Mezzanine on 3.4+.
This was an issue in the Ogre Samples build.  Additional checks were
being performed that depended on data that was not set regardless of the
samples being set to disabled.  Solution was to just comment out the
adding of the Samples subdirectory, as the Mezzanine has zero use for
that.