A number of changes made to the management of WorldObjects.
WorldObject factories on their managers used to be stored on each
instance of a given WorldObject manager.  This required the manager to
be constructed prior to the registration of a custom type.  Now they are
in static containers that can be populated whenever.  Also touched up
the Terrain classes a little and renamed TerrainBase to Terrain.