Fixed how WorldObjects were registered/unregistered.
The WorldObject factories were being populated in the constructor and
destroyed in the destructor.  This worked when factories were stored on
each instance, but that is no longer the case.  Registration was moved
to it's own method that is now called by the Entresol, and factories are
destroyed on Entresol destruction.  This avoids making duplicates (in
most situations) of factories.