Made a number of changes to how WorldProxies are managed.
Nothing earthshattering, but a couple of new interface classes for
WorldManagers have been introduced, one for WorldManagers that create
proxies, and one for those that create objects.  The Object interface is
still a WIP.  Proxy managers have been updated to conform to the new
interface.  The WorldProxy base class has been updated to return a
pointer of this interface when getting it's creator.  All WorldProxies
also must be able to say if they are static or not.  This is put into
place for longer term plans.  The TransformableObject interface used by
both proxies and objects gained a set of methods for using Transforms.