A fairly large refactor to the WorldObject class.
Rewrote the WorldObject class so it has an implementation for proxy
storage.  This cuts down on the number of methods that must be
overloaded dramatically (30 -> 2).  Also introduced the concept of a
Primary Proxy, which will be treated as central to the object and used
for queries that makes to only query one proxy.  Better Proxy management
is now built into the WorldObject class.  Additionally, Collisions are
no longer being reported to the WorldObject class directly.  The
collision system is still in place and can be accessed via the
PhysicsManager.  Just WorldObjects will no longer have that information
on them.