Made some changes to WorldObject update API and how transforms are sync'd from physics.
"_Update" on WorldObjects now takes a Whole representing the amount of
time elapsed since the last update.  Currently nothing is done with this
but in the future some WorldObjects will depend on it.  Additionally, in
the old system transforms were sent from the RigidProxy to all objects
that opted in.  Sync objects had to be explicitly added.  Now the
RigidProxy will only be aware of it's Parent Object when updating
transforms and will update every proxy in it other than itself.  As a
result of this the ability to follow another WorldObject was added to
the AreaEffect class.  Lastly, a small fix to supporting multiple
physicsmanagers was made by replacing a global with a variable.