Made a lots of small improvements to geometry classes and functionality.
We had been using types defined in the MathTools namespace for returns
of intersection tests on some but not all geometry classes.  I've
updated the list of types and moved them into a new file "mathtypes.h",
and then updated the remaining classes to use types listed in that
header.
We were also using Ray to represent a segment in a number of these
return types, so I changed the pairs that were using a Ray to use a
LineSegment3D instead and updated the appropriate sources.
At the core of the changes are the improvements to the Triangle3D class
which has new intersection testing methods.  This will enable us to rely
on Ogre math less when doing ray casts on meshes.  Two algorithms were
implemented, one which has early escapes for a known normal, and another
that doesn't need a normal.