Updated AssetID to reflect it's need to serialize stream targets.
AssetID was moved to the Mezz namespace, and was slightly reconfigured
and can be serialized.  Also added a couple of convenience methods to
the Enums file.  A method to get the n-th bit value, as well as
something to convert enum class values to ints (which is good for
serialization).