Bug fixes related to testing the RenderableRayQuery.
Fixed a number of issues with the RayQuery itself, including the setting
of wrong flags, and not selectively using parameters when we're supposed
to.  The Ogre::ManualObject used by the MeshGenerator was applying
padding we didn't ask for and couldn't readily configure to all meshes
generated.  A workaround was made to remove the padding.  The
BoxGenerator was treating the HalfExtents passed in as the entire
extents to be generated and has been fixed.