Made some minor tweaks and additions to the HashedString and SortedVector classes.
Updated the HashedString to have copy and move semantics.  Added global
operators to work with string literals.
In the SortedVector I changed "add" to return an iterator, and it uses
std::lower_bound to insert.  Also added an "add_emplace" that accepts a
compare method/functor/lambda and a variadic template for emplace
construction.