Small series of fixes for the LineGroupProxy to finally get it rendering.
I wrote a todo in the physicsmanager source file describing the issue
somewhat, but the LineGroupProxy was failing to render because we seem
to have been allocating buffers too early.  I have changed the code back
to how it worked pre-refactor (not all of it, just this tiny part) so
that it allocates a hardware buffer on demand instead of at the soonest
possible time.  This should be looked into at some point.