Added a converting constructor to the HashedString32.
Added a constructor that accepts a char* to the HashedString32, allowing
it to be more of a drop in replacement for String where we want it to
be.  Also added assignment operators for std::string and char*.
Doing this causing abiguity errors for the StateMachine, as it had
overloads that accepted String and HashedString32 datatypes.  I removed
the String overloads where applicable to remove the error, and they are
no longer necessary.