Some more minor tweaks to the Event system code.
Removed the underscores from the DispatchEvent method.  It's not really
dangerous enough to warrent it.  Renamed most things on the
EventPublisher that mentioned "EventTable" to "SubscriptionTable".  This
had the unfortunate effect of lengthing the method and type names.
Probably still the lesser of two evils.  Made use of std::make_shared to
create our bindings.