Overhauling the Input system with the primary focus of not relying on the event system.
The overall goal is for the Input system to perform the internal event
pump on it's own and access only the events it is interested in.  During
the course of this I've done touchups to every file in the Input
namespace, as well as made separate classes for Joystick and Controller
since SDL has a GameController API that I was unaware of.