Updated the Graphics system to use the new event system.
Created WindowEvent and WindowTransformEvent to graphics system dispatch
related to windows.  Updated GraphicsManager to pump the internal events
and dispatch them appropriately.  Did some code and documentation
cleanup while I was there.  Added a convenience method to the Viewport
that will check if a point is within it's bounds on the Window.  Added a
method that will retrieve the ID of the GameWindow, which is used with
the dispatched events.