Last round of touchups.
Did a refactor of the SequenceContainer and now the code is a little bit
cleaner.  More changes need to happen but they'll be later.  Changed a
viewport method to take two Wholes instead of a Vector2, which removed
the need for some casting.  Added <chrono> to the datatypes header (for
some future changes I have planned).