Incremental updates to the new event system.
I have split the EventSubscriptionTable into two classes.  The new EventBindingTable also has a couple of helper classes in it's source files that help it seemlessly interact with other support classes in the event system.  I created a new EventID class, as well as a tiny namespace with some helper conversion methods.  Numerous fixes and updates to documenation as well.