Fixed an issue where the reference type was wrong when passed between template methods.
The issue was in the FunctionSubscriber class, which had defined the parameters as a part of it's type, rather than on the method to be deduced.  Also changed around how types were passed in and it can now support return values if needed.

Enabled perfect forwarding where I could in all other dispatch methods.