A large refactor of the EventSubscriptionTable.
EventSubscriptionTable inherits from a couple of other event classes now in order to establish it's API.  There are more support classes for the centralization and passing around of necessary types which can be configured with a couple easy steps.  The event system is pretty close to it's final form, although it requires more UnitTests.  The current set of UnitTests all pass.