Another round of updates to the threading system.
Suspiciously, this repo was building for C++11, not 14.  That has been corrected.  Added features for deducing types in member functions.  Updated the Dispatchers a great deal, including the ability to get return values from functions that are dispatched to.  The only current exception to this is getting return values from the Queued Dispatcher, as that requires some special features.  Added "HasSubscription" to all the container classes.  Removed some commented out code.  Added many more UnitTests.  The event system is up to 95 tests.