Another round of updates to the event system
Purged the EventBindingTable.  That functionality is entirely handled by the binding factory now.  Added more tests.  The only tests missing now are the Query tests for the QueuedDispatcher.  Implemented the move constructor/assignment operator on the ManagedArray to fix some memory corruption issue.  Cleaned up some types that didn't need to exist, and reduce copying in the system.  Fleshed out the remaining functionality in the Single sub container.