Commiting all updates to the resource system thus far.
This is a large amount of changes to DataStreams and the Resource system, with a couple tweaks to utilities used by the resource system.  This contains the initial version of archive readers and writers for zip and filesystem(7z pending).  The resource utilities mega file has been split into several files with a narrower focus.  Datastream API has been updated to better reflect the needs of archives, which didn't take much.  The older archive monolith classes didn't match all the design requirements of some of the crazier archive formats, and thus were removed having the additional benefit of simplifying their design.